feat(suite-common): move getTokens utility and token categorisation selectors to suite-common
Move getTokens() from the desktop-only tokenUtils and introduce new memoized
selectors (selectAccountTokens, selectAccountHiddenTokens, selectAccountDefiTokens,
selectAccountDefiTokensCount) in suite-common/wallet-core so they can be shared
across desktop and native.
